Added more comments in my part of the code
After conquering the seas for a decade, Captain Longneck went back to his ancestral earth to spend the last of his days... And so ends the tale of the tallest pirate to ever roam the seven seas. But, while Captain Longneck may have died, his heritage lives on. We will always remember you, captain... Farewell.